requirements.txt
bcrypt==3.2.0
blinker==1.8.2
cffi==1.14.5
click==8.1.7
dnspython==2.1.0
email-validator==1.1.2
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy==1.11.1
mypy-extensions==1.0.0
numpy==2.1.0
Pillow==8.2.0
pycparser==2.20
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==2.3.3